Add API to start new sessions

+++ moth_radio/models.py
@@ -19,8 +19,8 @@ class Session(db.Model):
 	id = db.Column(db.Integer, primary_key= True, unique=True)
 	labUserId = db.Column(db.Integer, db.ForeignKey('labUsers.id'))
 	psiturkUid = db.Column(db.String(64))
-	startTime = db.Column(db.Float)
-	stopTime = db.Column(db.Float)
+	startTime = db.Column(db.Integer)
+	stopTime = db.Column(db.Integer)
 
 class Rating(db.Model):
 	__tablename__ = 'ratings'
